final_work/requirements.txt
pytest==9.0.2
playwright==1.58.0
allure-pytest==2.15.3
flake8
